update projection manager and add abstract projection manager test
